% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcFormula}
\alias{calcFormula}
\title{Calculate peptide formulas}
\usage{
calcFormula(sequences, subscripts = FALSE, residueAtoms = "")
}
\arguments{
\item{sequences}{Peptide sequences}

\item{subscripts}{Should formulas have subscripts or normal baseline numbers?}

\item{residueAtoms}{Path to residueAtoms file. If blank, the default file included in the package is used.}
}
\value{
vector of peptide formulas.
}
\description{
Calculate peptide molecular formulas
}
\examples{
calcFormula(c("ACLLPETVNMEEYPYDAEY", "ALCAEFK"), subscripts = TRUE)

}
