% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcMass}
\alias{calcMass}
\title{Calculate peptide masses}
\usage{
calcMass(sequences, monoMass = TRUE, residueAtoms = "",
  atomMasses = "")
}
\arguments{
\item{sequences}{Peptide sequences}

\item{monoMass}{Should monoisotopic mass be calculated. If false, average mass is calculated.}

\item{residueAtoms}{Path to residueAtoms file. If blank, the default file included in the package is used.}

\item{atomMasses}{Path to atomMasses file. If blank, the default file included in the package is used.}
}
\value{
vector of peptide masses.
}
\description{
Calculate peptide monoisotopic or average masses.
}
