% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combineMods}
\alias{combineMods}
\title{Combined mods from multiple peptides into a single string}
\usage{
combineMods(mods, sep = "|")
}
\arguments{
\item{mods}{Modifications to combine}

\item{sep}{delimiter separating modifications}
}
\value{
Modifications combined into a single string
}
\description{
Combined concated mods from multiple peptides into a single string
}
