% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{digest}
\alias{digest}
\title{Perform a virtual protease digest of a protein.}
\usage{
digest(sequences, ids, nMissedCleavages = 0L,
  cleavagePattern = "([RK])(?=[^P])", mz_filter = TRUE,
  residueAtoms = "", minMz = 400, maxMz = 1800, minCharge = 1L,
  maxCharge = 5L, minLen = 6L, maxLen = 0L)
}
\arguments{
\item{sequences}{StringVector containing protein sequences. Whitespace will automatically be removed.}

\item{ids}{Names for the slot for each protein's peptides in output.}

\item{nMissedCleavages}{number of missed cleavages to allow.}

\item{cleavagePattern}{RegEx for protease cleavage pattern. The default is the pattern for trypsin.}

\item{mz_filter}{Should peptides included in output be filtered by mz?}

\item{residueAtoms}{Path to residueAtoms file. If blank, the default file included in the package is used.}

\item{minMz}{Minimum m/z to allow in peptides.}

\item{maxMz}{Maximum m/z to allow in peptides. Set to 0 for no upper bound on m/z.}

\item{minCharge}{Minimum charge to consider when calculating m/z.}

\item{maxCharge}{Maximum charge to consider when calculating m/z.}

\item{minLen}{Minimum peptide length.}

\item{maxLen}{Maximum peptide length. Set to 0 for no upper bound on length.}
}
\value{
A list with named elements containing vectors of each input protein's peptides.
}
\description{
The function uses charge and m/z filters to remove peptides which would not be
observable by MS. The m/z for peptides in charge states minCharge to maxCharge
are calculated. If the m/z for any charge state is in between minMZ and maxMZ, the
sequence will be appended to peptides.
}
\examples{
digest(c("KLGAARKLGAGLAKVIGAGIGIGK", "KLGAARKLGAGLAKPVIGAGIGIGK"), c('a', 'b'))

}
