% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastaInfo}
\alias{fastaInfo}
\title{Get fasta file info.}
\usage{
fastaInfo(fastaPath = "")
}
\arguments{
\item{fastaPath}{Path to fasta file. Be default, fasta file included in package is used.}
}
\value{
List with slots for sequence count and vector of entry IDs contained in file.
}
\description{
Get metadata about a fasta file.
}
