% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_uniprot_sequences}
\alias{fetch_uniprot_sequences}
\title{Download protein sequences from the UniProt API for a list of accessions.}
\usage{
fetch_uniprot_sequences(
  accessions,
  output_file = NULL,
  use_cached = FALSE,
  batch_size = 100
)
}
\arguments{
\item{accessions}{A vector of UniProt accessions to download.}

\item{output_file}{An optional file path to a fasta file to write sequences to.}

\item{use_cached}{If TRUE, read and return the matching ids in \code{output_file} if the file exists.
Otherwise download the sequences through the UniProt API.
An error is raised if any accession is not found in the cached file.}

\item{batch_size}{Number of accessions to fetch per API request (default 100)}
}
\value{
A data.frame with columns \code{id} and \code{sequence},
        matching the format of \code{\link{readFasta}}.
}
\description{
Download protein sequences from the UniProt API for a list of accessions.
}
\examples{
fetch_uniprot_sequences(c('P00533'))

}
