% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFragmentIonIndices}
\alias{getFragmentIonIndices}
\title{Get fragment ion sequence indices.}
\usage{
getFragmentIonIndices(seq)
}
\arguments{
\item{seq}{Peptide sequence as a string.}
}
\value{
fragments IntegerMatrix where each row is a fragment ion. The first column is the startig index of the fragment ion in the peptide sequence and the second column is the fragment length.
}
\description{
Get indices of fragment ions for peptide sequence.
}
