% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFragmentIonSequences}
\alias{getFragmentIonSequences}
\title{Get fragment ion sequences.}
\usage{
getFragmentIonSequences(seq)
}
\arguments{
\item{seq}{Peptide sequence as a string.}
}
\value{
fragments Named CharacterVector where the names are the ions and the values are the sequences.
}
\description{
Get sequences of fragment ions for peptide sequence.
}
