% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMass}
\alias{getMass}
\title{Calculate peptide masses}
\usage{
getMass(sequences, monoMass = TRUE, residueAtoms = "",
  atomMasses = "")
}
\arguments{
\item{sequences}{Peptide sequences}

\item{monoMass}{Should monoisotopic mass be calculated. If false, average mass is calculated.}

\item{atomMasses}{Path to atomMasses. If blank, the default file included in the package is used.}

\item{residueMasses}{Path to residueAtoms file. If blank, the default file included in the package is used.}
}
\value{
vector of peptide masses.
}
\description{
Calculate peptide monoisotopic or average masses.
}
