% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getModifiedResidues}
\alias{getModifiedResidues}
\title{Get locations of modified residues in parent protein}
\usage{
getModifiedResidues(ids, peptideSeq, fastaPath = "", modSep = "|")
}
\arguments{
\item{ids}{CharacterVector of Uniprot IDs}

\item{peptideSeq}{CharacterVector of peptide sequences containing modifications}

\item{fastaPath}{path to fasta formated file to look up protein sequences}

\item{modSep}{delimiter for multiple modifications}
}
\value{
CharacterVector containing locations of modifications in protein sequence
}
\description{
Get locations of modified residues in parent protein
}
\examples{
getModifiedResidues(c("Q00839", "Q9HCS7", "Q7L014"), c("APQC*LGK", "FADMEC*K", "GAEIIVC*TPGR"))

}
