% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMSScanMetadata}
\alias{getMSScanMetadata}
\title{Read MS scan metadata from mzML files.}
\usage{
getMSScanMetadata(scans, files, progressBar = TRUE, nThread = 0L)
}
\arguments{
\item{scans}{IntegerVector of scan numbers.}

\item{files}{CharacterVector of file paths to mzML files. Must be the same length as scans.}

\item{progressBar}{Show progress bar?}

\item{nThread}{Number of threads to use. By default use 1 thread per virtual core on machine.}
}
\value{
DataFrame with columns for scan, file, path, rt, ms_level,
  precursor_mz, isolation_window_lower_offset, and isolation_window_upper_offset.
}
\description{
Read scan metadata from mzML files.
}
\examples{
mzml_path <- system.file('extdata/20240927_STEL1_Evo3_MW_15N_MS3_dilution_12_1.mzML',
                          package = 'RPeptideUtils')
getMSScanMetadata(c(1L, 2L, 3L), rep(mzml_path, 3), progressBar = FALSE)

}
