% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{H1_MASS}
\alias{H1_MASS}
\title{Monoisotopic mass of hydrogen (H1)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
H1_MASS
}
\description{
Monoisotopic mass of hydrogen (H1)
}
\keyword{datasets}
