% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indexN}
\alias{indexN}
\title{get the index of nth residues of query in ref.}
\usage{
indexN(query, ref, n = 1L, noExcept = FALSE)
}
\arguments{
\item{query}{String to search for.}

\item{ref}{String to search in.}

\item{n}{Residue number in query.}

\item{noExcept}{Should an std::runtime_error be thrown if query is not in ref?}
}
\value{
n residues after query.
}
\description{
Get the index of residue n of query in ref. If n is -1, the index of the last residue in query is returned.
}
