% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matchingProteins}
\alias{matchingProteins}
\title{Find all proteins containing peptide sequences.}
\usage{
matchingProteins(peptides, fastaPath = "", progressBar = TRUE, nThread = 0L)
}
\arguments{
\item{peptides}{A character vector of peptide sequences.}

\item{fastaPath}{path to fasta formated file to look up protein sequences.}

\item{progressBar}{Show progress bar?}

\item{nThread}{Number of threads to use. By default use 1 thread per virtual core on machine.}
}
\value{
List where names are peptide sequences, and values are the ids for the matching proteins.
}
\description{
Given a vector of peptide sequences, find all the proteins containing the peptied in a fasta file.
}
\examples{
fasta_path <- system.file('extdata/Human_uniprot-reviewed_20171020.fasta', package = 'RPeptideUtils')
matchingProteins(c('PEPTIDE'), progressBar = FALSE, fastaPath = fasta_path)

}
