% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nAfter}
\alias{nAfter}
\title{get n residues after query.}
\usage{
nAfter(query, ref, n, noExcept = FALSE)
}
\arguments{
\item{query}{String to search for.}

\item{ref}{String to search in.}

\item{n}{Number of residues in output.}

\item{noExcept}{Should an std::runtime_error be thrown if query is not in ref?}
}
\value{
n residues after query.
}
\description{
Get n residues after query in ref. If n overruns ref, the maximum number of characters will be returned.
}
