% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nBefore}
\alias{nBefore}
\title{get n residues before query.}
\usage{
nBefore(query, ref, n, noExcept = FALSE)
}
\arguments{
\item{query}{String to search for.}

\item{ref}{String to search in.}

\item{n}{Number of residues in output.}

\item{noExcept}{Should an std::runtime_error be thrown if query is not in ref?}
}
\value{
n residues before query.
}
\description{
Get n residues before query in ref. If n overruns ref, the maximum number of characters will be returned.
}
