% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nMissed}
\alias{nMissed}
\title{Count missed protease cleavages}
\usage{
nMissed(sequences, cleavagePattern = "([RK])(?=[^P])")
}
\arguments{
\item{sequences}{CharacterVector of peptide sequences.}

\item{cleavagePattern}{Regex pattern for protease cleavage sites. The default is the pattern for trypsin.}
}
\value{
IntegerVector of missed cleavage counts.
}
\description{
Count the number of missed protease cleavages in peptide sequences.
}
\examples{
nMissed(c("PEPTIDER", "PEPKTIDER", "PEPKTIDERKK"))

}
