% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oneLetterToThree}
\alias{oneLetterToThree}
\title{Convert to 3 letter amino acid codes}
\usage{
oneLetterToThree(sequences, sep_in = "", sep_out = "",
  n_term_out = "", c_term_out = "")
}
\arguments{
\item{sequences}{vector of sequences}

\item{sep_in}{deliminator between amino acids in input}

\item{sep_out}{deliminator between amino acids in output}

\item{n_term_out}{string to append to n terminus}

\item{c_term_out}{string to append to c terminus}
}
\value{
StringVector of peptides with three letter amino acid codes
}
\description{
Convert from 1 letter amino acid codes to 3
}
