\name{peptideUtils-package}
\alias{peptideUtils-package}
\alias{peptideUtils}
\docType{package}
\title{
  R package which wraps useful c++ functions for dealing with peptide data.
}
\description{
  This package wraps useful functions from my utils library (https://github.com/ajmaurais/utils) for working with string representations of peptides in R.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Aaron Maurais

Maintainer: Aaron Maurais mauraisa@bc.edu
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
