% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROTON_MASS}
\alias{PROTON_MASS}
\title{Mass of a proton (H+)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PROTON_MASS
}
\description{
Use this constant (not \code{H1_MASS}) when calculating m/z from neutral mass:
\code{mz = (M + z * PROTON_MASS) / z}
}
\keyword{datasets}
