% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readFasta}
\alias{readFasta}
\title{Read fasta file.}
\usage{
readFasta(fastaPath = "", n_entries = 0L)
}
\arguments{
\item{fastaPath}{Path to fasta file. Be default, fasta file included in package is used.}

\item{n_entries}{Number of entries to read. If 0, all entires are read.}
}
\value{
DataFrame with columns for ID and sequence.
}
\description{
Read all sequences in fasta file. Reverse matches are automatically skipped.
}
