% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threeLetterToOne}
\alias{threeLetterToOne}
\title{Convert to 1 letter amino acid codes}
\usage{
threeLetterToOne(
  sequences,
  sep_in = "",
  sep_out = "",
  n_term_out = "",
  c_term_out = ""
)
}
\arguments{
\item{sequences}{vector of sequences}

\item{sep_in}{deliminator between amino acids in input}

\item{sep_out}{deliminator between amino acids in output}

\item{n_term_out}{string to append to n terminus}

\item{c_term_out}{string to append to c terminus}
}
\value{
StringVector of peptides with one letter amino acid codes
}
\description{
Convert from 3 letter amino acid codes to 1
}
\examples{
threeLetterToOne(c("Ala-Cys*-Leu-Leu-Pro", "Ala-Leu-Cys-Ala", "Ala-Gln-Sec-Ile"), sep_in = "-")

}
