% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transpose_sequence}
\alias{transpose_sequence}
\title{Transpose peptide quantifications into amino acid level quantifications.}
\usage{
transpose_sequence(peptide_sequences, quantification, protein_seq)
}
\arguments{
\item{peptide_sequences}{List of peptide sequences.}

\item{quantification}{Ratio or spectral count values for peptide_sequences.}

\item{protein_seq}{Parent protein sequence.}
}
\value{
DataFrame with columns for 'residue', 'number', and 'quant'.
}
\description{
Transpose peptide quantifications for a single protein into amino acid level
quantifications. An row will be included in the output for each time an amino acid
at a given position was included in a peptide in peptide_sequences. Additional processing
is required to obtain summary values for each amino acid position.
}
