% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combineMods}
\alias{combineMods}
\title{Combine mods from multiple peptides into a single string}
\usage{
combineMods(mods, sep = "|")
}
\arguments{
\item{mods}{Modifications to combine}

\item{sep}{delimiter separating modifications}
}
\value{
Modifications combined into a single string
}
\description{
Combine concated mods from multiple peptides into a single string.
}
\examples{
combineMods(c('C157', 'C157|C125', 'C50', 'C125'))

}
