% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getModifiedResidues}
\alias{getModifiedResidues}
\title{Get locations of modified residues in parent protein}
\usage{
getModifiedResidues(fastaPath, ids, peptideSeq, modSep = "|")
}
\arguments{
\item{fastaPath}{path to fasta formated file to look up protein sequences}

\item{ids}{CharacterVector of Uniprot IDs}

\item{peptideSeq}{list of peptide sequences containing modifications}

\item{modSep}{delimiter for multiple modifications}
}
\value{
CharacterVector containing locations of modifications in protein sequence
}
\description{
Get locations of modified residues in parent protein
}
