% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSquences}
\alias{getSquences}
\title{Get protein IDs in fasta file for a vector of Uniprot IDs}
\usage{
getSquences(fastaPath, ids)
}
\arguments{
\item{fastaPath}{path to fasta formated file to look up protein sequences}

\item{ids}{CharacterVector of uniprot IDs}
}
\value{
CharacterVector of protein sequences in same order as ids.
}
\description{
Get protein sequences for a vector of uniprot IDs
}
