\name{peptideUtils-package}
\alias{peptideUtils-package}
\alias{peptideUtils}
\docType{package}
\title{
  An R package for working with string representations of peptides
}
\description{
  This package wraps useful c++ functions from my utils library (https://github.com/ajmaurais/utils) for working with string representations of peptides in R.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Aaron Maurais

Maintainer: Aaron Maurais mauraisa@bc.edu
}
