% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ids}
\alias{read_ids}
\title{Read a list of IDs, and protein names from a fasta file.}
\usage{
read_ids(
  fname = system.file("extdata/Human_uniprot-reviewed_20171020.fasta", package =
    "peptideUtils", mustWork = T),
  id_line_regex = "^>(sp|tr)",
  capture_regex = "^>[sptr]{2}\\\\|(\\\\w+)\\\\|(\\\\w+)_\\\\w+",
  .colnames = c("id", "protein")
)
}
\arguments{
\item{fname}{Path to fasta file to get data from.}

\item{id_line_regex}{Regex to find fasta entry header lines}

\item{capture_regex}{Regex with 2 capturing groups.
The first will be put in the \code{id} column, the second will be put in the \code{protein}.}

\item{.colnames}{Character vector of length 2 containing the column names in the output.}
}
\description{
Read a list of IDs, and protein names from a fasta file.
}
\examples{
read_ids()

}
