% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atomCountTable.R
\name{write_atomCountTable}
\alias{write_atomCountTable}
\title{Write default atomCountTable to current working directory}
\usage{
write_atomCountTable(path = getwd(), overwrite = F)
}
\arguments{
\item{path}{file path to write to}
}
\value{
TRUE is sucsssful
}
\description{
Write default atomCountTable to current working directory
}
